Keep collector internals file-local
