+++ projects/rocprofiler-compute/src/lib/torch_trace_collector/process_state.h
@@ -1,9 +1,7 @@
 // Copyright (c) Advanced Micro Devices, Inc.
 // SPDX-License-Identifier:  MIT
 //
-// The collector's process-wide state. at::addGlobalCallback registers one
-// callback pair for the whole process, and RecordFunction passes the callbacks
-// no state of its own, so they reach it through process_state().
+// Process-wide collector state.
 
 #pragma once
 
@@ -18,14 +16,11 @@ namespace torch_trace_collector::detail
 
 using rocprofiler_compute_tool::common::synchronized_t;
 
-// Records the RecordFunction callback registration.
 struct InstallState
 {
-    at::CallbackHandle handle    = at::INVALID_CALLBACK_HANDLE;
-    bool               installed = false;
+    at::CallbackHandle handle = at::INVALID_CALLBACK_HANDLE;
 };
 
-// Per-thread state is separate; see ThreadState in marker_stack.h.
 struct ProcessState
 {
     Stats                        stats;
@@ -33,8 +28,6 @@ struct ProcessState
     SnapshotStore                snapshots{stats};
 };
 
-// The one instance, constructed on first use. Each binary that links the
-// collector holds its own and registers its own callback.
 ProcessState& process_state();
 
 }  // namespace torch_trace_collector::detail
