+++ projects/rocprofiler-compute/src/lib/torch_trace_collector/torch_trace_collector_module.cpp
@@ -5,8 +5,6 @@
 
 #include "process_state.h"
 #include "record_function_installation.h"
-#include "snapshot_store.h"
-#include "stats.h"
 #include "user_scope.h"
 
 #include <pybind11/pybind11.h>
@@ -48,12 +46,13 @@ PYBIND11_MODULE(torch_trace_collector, m)
     m.def("install", &install, "Install the global RecordFunction callback. Idempotent.");
     m.def("uninstall", &uninstall, "Remove the registered callback.");
     m.def("is_installed", &is_installed, "Return True if the callback is installed.");
-    m.def("push_user_scope",
-          &push_user_scope,
-          pybind11::arg("marker"),
-          pybind11::arg("context"),
-          pybind11::arg("backend") = std::string(""),
-          "Push a USER_SCOPE frame, emit a ROCTX range, publish chain into TLS DebugInfo.");
-    m.def("pop_user_scope", &pop_user_scope, "Pop the most recent push_user_scope() frame on this thread.");
-    m.def("dump_stats", &dump_stats, "Internal counters for tests/debugging.");
+    m.def(
+        "push_user_scope",
+        &push_user_scope,
+        pybind11::arg("marker"),
+        pybind11::arg("context"),
+        pybind11::arg("backend") = std::string(""),
+        "Push a marker frame, emit a ROCTX range, and publish the stack to ThreadLocalDebugInfo.");
+    m.def("pop_user_scope", &pop_user_scope, "Pop the most recent push_user_scope frame on this thread.");
+    m.def("dump_stats", &dump_stats, "Return collector counters.");
 }
