+++ projects/rocprofiler-compute/src/lib/torch_trace_collector/user_scope.h
@@ -3,40 +3,16 @@
 
 #pragma once
 
-#include "stack_entry.h"
-
-#include <c10/util/ThreadLocalDebugInfo.h>
-
-#include <cstddef>
 #include <string>
-#include <utility>
-#include <vector>
 
 namespace torch_trace_collector::detail
 {
 
-// Carries the main-thread USER_SCOPE chain to autograd workers.
-class RoctxUserScopeChain : public c10::DebugInfoBase
-{
-public:
-    explicit RoctxUserScopeChain(std::vector<StackEntry> c)
-        : chain(std::move(c))
-    {
-    }
-
-    std::vector<StackEntry> chain;
-};
-
-// ThreadLocalDebugInfo kind for the USER_SCOPE chain.
-extern const c10::DebugInfoKind kRoctxUserScopeKind;
-
-// Overlays the published USER_SCOPE chain onto the thread stack.
-std::size_t apply_userscope_overlay();
-
-// Pushes a USER_SCOPE frame and emits a ROCTX range. When non-empty,
-// backend is appended to the range as "|<backend>".
+// Pushes a marker frame, publishes the thread stack to ThreadLocalDebugInfo,
+// and emits a ROCTX range. A non-empty backend is appended as "|<backend>".
 void push_user_scope(const std::string& marker, const std::string& context, const std::string& backend);
 
+// Pops the matching marker frame and ROCTX range.
 void pop_user_scope();
 
 }  // namespace torch_trace_collector::detail
