+++ projects/rocprofiler-compute/src/utils/inject_roctx/registry.py
@@ -24,7 +24,7 @@ _REGISTRY: dict[str, Backend] = {}
 
 
 def register(backend: Backend) -> None:
-    """Register backend under backend.name."""
+    """Register a backend."""
     _REGISTRY[backend.name] = backend
 
 
@@ -32,11 +32,7 @@ def install_many(names: Iterable[str]) -> None:
     """Install each named backend.
 
     Unknown or failing backends are skipped after a warning.
-    ``UnsupportedTorchVersionError`` is re-raised.
     """
-    from utils.inject_roctx._backends.torch_cpp_loader import (
-        UnsupportedTorchVersionError,
-    )
     from utils.logger import console_warning
 
     seen: set[str] = set()
@@ -53,8 +49,6 @@ def install_many(names: Iterable[str]) -> None:
                     f"but did not register {name!r}"
                 )
             _REGISTRY[name].install()
-        except UnsupportedTorchVersionError:
-            raise
         except Exception as exc:
             console_warning(
                 "inject_roctx",
